# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Main Negotiation Bot API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/dashboard": {
   "get": {
    "summary": "Dashboard",
    "operationId": "dashboard_api_dashboard_get",
    "parameters": [
     {
      "name": "date_from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Date From"
      }
     },
     {
      "name": "date_to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Date To"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Dashboard"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/events": {
   "get": {
    "summary": "Events",
    "operationId": "events_api_events_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "direction",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "buy",
          "sell"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Direction"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "received",
          "in_progress",
          "closed"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "category_key",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Category Key"
      }
     },
     {
      "name": "date_from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Date From"
      }
     },
     {
      "name": "date_to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Date To"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/EventView"
         },
         "title": "Response Events Api Events Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Add Event",
    "operationId": "add_event_api_events_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NewEvent"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/vendor-suggestions": {
   "get": {
    "summary": "Vendor Suggestions",
    "operationId": "vendor_suggestions_api_vendor_suggestions_get",
    "parameters": [
     {
      "name": "direction",
      "in": "query",
      "required": true,
      "schema": {
       "enum": [
        "buy",
        "sell"
       ],
       "type": "string",
       "title": "Direction"
      }
     },
     {
      "name": "category_key",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Category Key"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "default": [],
       "title": "Q"
      }
     },
     {
      "name": "category_label",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Category Label"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/handoff": {
   "post": {
    "summary": "Open Ais Case",
    "description": "Open (or reopen) the negotiation for a case that lives in the AIS prototype.",
    "operationId": "open_ais_case_api_handoff_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/HandoffIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HandoffOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/events/simulate": {
   "post": {
    "summary": "Simulate Event",
    "operationId": "simulate_event_api_events_simulate_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SimulateIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/events/{event_id}": {
   "get": {
    "summary": "Event",
    "operationId": "event_api_events__event_id__get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/items": {
   "get": {
    "summary": "Items",
    "operationId": "items_api_items_get",
    "parameters": [
     {
      "name": "event_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Event Id"
      }
     },
     {
      "name": "has_bids",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Has Bids"
      }
     },
     {
      "name": "recommendation",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "waiting",
          "negotiate",
          "accept",
          "review",
          "done"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Recommendation"
      }
     },
     {
      "name": "direction",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "buy",
          "sell"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Direction"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "date_from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Date From"
      }
     },
     {
      "name": "date_to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Date To"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ItemRow"
         },
         "title": "Response Items Api Items Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/history": {
   "get": {
    "summary": "History",
    "operationId": "history_api_history_get",
    "parameters": [
     {
      "name": "direction",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "buy",
          "sell"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Direction"
      }
     },
     {
      "name": "category_key",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Category Key"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "negotiated",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Negotiated"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 500,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/HistoryRow"
         },
         "title": "Response History Api History Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/items/{item_id}": {
   "get": {
    "summary": "Item",
    "operationId": "item_api_items__item_id__get",
    "parameters": [
     {
      "name": "item_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Item Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ItemDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/items/{item_id}/comparison": {
   "get": {
    "summary": "Item Comparison",
    "operationId": "item_comparison_api_items__item_id__comparison_get",
    "parameters": [
     {
      "name": "item_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Item Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ComparisonView"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/history/{record_id}": {
   "get": {
    "summary": "History Deal",
    "operationId": "history_deal_api_history__record_id__get",
    "parameters": [
     {
      "name": "record_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Record Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HistoryDeal"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/items/{item_id}/history": {
   "get": {
    "summary": "Item History",
    "operationId": "item_history_api_items__item_id__history_get",
    "parameters": [
     {
      "name": "item_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Item Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HistoryView"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/items/{item_id}/points": {
   "put": {
    "summary": "Set Points",
    "operationId": "set_points_api_items__item_id__points_put",
    "parameters": [
     {
      "name": "item_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Item Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PointsIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ItemDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Dashboard": {
    "properties": {
     "kpis": {
      "$ref": "#/components/schemas/Kpis"
     },
     "events": {
      "items": {
       "$ref": "#/components/schemas/EventView"
      },
      "type": "array",
      "title": "Events"
     },
     "value_by_category": {
      "items": {
       "$ref": "#/components/schemas/CategoryValue"
      },
      "type": "array",
      "title": "Value By Category"
     },
     "top_vendors": {
      "items": {
       "$ref": "#/components/schemas/VendorValue"
      },
      "type": "array",
      "title": "Top Vendors"
     },
     "opportunities": {
      "items": {
       "$ref": "#/components/schemas/Opportunity"
      },
      "type": "array",
      "title": "Opportunities"
     },
     "status_distribution": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Status Distribution"
     },
     "item_state_distribution": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Item State Distribution"
     },
     "delta_generated": {
      "$ref": "#/components/schemas/DeltaGenerated"
     },
     "insight": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Insight"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "kpis",
     "events",
     "value_by_category",
     "top_vendors",
     "opportunities",
     "status_distribution",
     "item_state_distribution",
     "delta_generated",
     "insight"
    ],
    "title": "Dashboard"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
  }
 }
}